+++ sandbox/wsgi.py
@@ -1,0 +1,13 @@
+"""
+WSGI config for sandbox.
+
+It exposes the WSGI callable as a module-level variable named ``application``.
+"""
+
+import os
+
+from django.core.wsgi import get_wsgi_application
+
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "settings.demo")
+
+application = get_wsgi_application()